requirements.txt
absl-py==1.1.0
addict==2.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
audioread==3.0.1
autopep8==2.0.4
av==11.0.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2021.5.30

coloredlogs==15.0.1
conda==4.10.3

cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
einops==0.7.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.33.3
fsspec==2023.12.2
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
huggingface-hub==0.20.1
humanfriendly==10.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata==7.0.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
lazy-loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4

mmcv==2.1.0
mmcv-full==1.7.1
mmengine==0.10.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
notebook-shim==0.1.0
numba==0.58.1
numpy==1.22.4
oauthlib==3.2.0
onnx==1.15.0
onnxruntime==1.16.3
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==4.1.0
pooch==1.8.0
prettytable==3.9.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.25.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycosat==0.6.3
pygments==2.17.2
pyparsing==3.0.9
pyrsistent==0.18.1

python-dateutil==2.8.2
pytorchtools==0.0.2
PyTurboJPEG==1.7.3
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==23.2.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.8
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.0
shapely==2.0.2
sniffio==1.2.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
stack-data==0.3.0
supervisor==4.2.4
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminado==0.15.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
traitlets==5.3.0
typing-extensions==4.2.0
tzdata==2023.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
yapf==0.40.2
zipp==3.8.0
